pyproject.toml
[tool.poetry]
name = "code-jam-2023"
version = "0.1.0"
description = "project made for python discord code jam 10"
authors = [
    "p0lygun <solankivibhakar82@gmail.com>",
    "Hamza-Farrukh <hamzafarrukh73@gmail.com>",
    "Anas-bot <anasihabezmc@gmail.com>",
    "Kaezrr",
    "rudravashishtha"
]
license = "MIT"
readme = "README.md"
packages = [{include = "game"}]

[tool.poetry.dependencies]
python = "^3.11"
numpy = "^1.25.2"
python-dotenv = "^1.0.0"
pygame-ce = "^2.3.1"
pygame-gui = "^0.6.9"
pillow = "^10.0.0"
opencv-contrib-python = "^4.8.0.76"

[tool.poetry.group.dev.dependencies]
pre-commit = "^3.3.3"
flake8 = "^6.1.0"
isort = "^5.12.0"
flake8-docstrings = "^1.7.0"

[tool.poetry.group.extras]
optional = true


[tool.poetry.group.extras.dependencies]
requests = "^2.31.0"
beautifulsoup4 = "^4.12.2"
svg-path = "^6.3"
svgpathtools = "^1.6.1"
lxml = "^4.9.3"
pycountry = "^22.3.5"
pyopengl = "^3.1.7"


[tool.poetry.scripts]
game = "game.__main__:run_game"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
